+++ classes/jigoshop_tax.class.php
@@ -307,7 +307,7 @@ class jigoshop_tax extends Jigoshop_Base {
 		if($allowed_countries === 'specific'){
 			$specific_countries = Jigoshop_Base::get_options()->get_option('jigoshop_specific_allowed_countries');
 			$base_cc = jigoshop_countries::get_base_country();
-			if(!in_array($country, $specific_countries)){
+			if(is_array($specific_countries) && !in_array($country, $specific_countries)){
 				if(in_array($base_cc, $specific_countries)){
 					$country = $base_cc;
 				} else {
@@ -369,7 +369,7 @@ class jigoshop_tax extends Jigoshop_Base {
 
 		if($allowed_countries === 'specific'){
 			$specific_countries = Jigoshop_Base::get_options()->get_option('jigoshop_specific_allowed_countries');
-			if(!in_array($country, $specific_countries)){
+			if(is_array($specific_countries) && !in_array($country, $specific_countries)){
 				$country = array_shift($specific_countries);
 			}
 			if(!in_array($state, array_keys($this->rates[$country]))){
@@ -756,7 +756,7 @@ class jigoshop_tax extends Jigoshop_Base {
 
 		if($allowed_countries === 'specific'){
 			$specific_countries = Jigoshop_Base::get_options()->get_option('jigoshop_specific_allowed_countries');
-			if(!in_array($country, $specific_countries)){
+			if(is_array($specific_countries) && !in_array($country, $specific_countries)){
 				$country = array_shift($specific_countries);
 			}
 		}